violations/api/violations_router.py
from fastapi import APIRouter, HTTPException
import logging

router = APIRouter()
logger = logging.getLogger(__name__)

@router.get("/bbl/{bbl}")
def get_violations(bbl: str):
    """Get NYC DOB violations by BBL"""
    try:
        # TODO: Import your dob_engine
        # from core.dob_engine import fetch_violations
        # violations = fetch_violations(bbl)
        
        # Mock response for now
        violations = [
            {
                "violation_number": "V123456",
                "description": "Failure to maintain building facade",
                "date": "2024-01-15",
                "status": "Open"
            }
        ]
        
        return {
            "bbl": bbl,
            "count": len(violations),
            "violations": violations
        }
    except Exception as e:
        logger.error(f"Error fetching violations: {str(e)}")
        raise HTTPException(500, f"Failed to fetch violations: {str(e)}")

@router.get("/search")
def search_violations(address: str):
    """Search violations by address"""
    # TODO: Implement address search
    return {
        "address": address,
        "message": "Search functionality coming soon"
    }
